# alora-auth-api/docs/swagger.yaml
basePath: /
definitions:
...
  CentralTokenResponse:
    properties:
      access_token:
        example: eyJhbGciOiJSUzI1NiIsImtpZCI6ImsxIiwidHlwIjoiYXQrand0In0...
        type: string
      expires_in:
        example: 900
        type: integer
      token_type:
        example: Bearer
        type: string
    type: object
...
  ChooseCompanyRequest:
    properties:
      client_id:
        example: 5c6d7e8f-9a0b-1c2d-3e4f-5a6b7c8d9e0f
        type: string
    required:
    - client_id
    type: object
...
  CompanyChoice:
    properties:
      client_id:
        example: 5c6d7e8f-9a0b-1c2d-3e4f-5a6b7c8d9e0f
        type: string
      name:
        example: Acme Corp
        type: string
    type: object
  CompanyChoices:
    properties:
      companies:
        items:
          $ref: '#/definitions/CompanyChoice'
        type: array
    type: object
...
  ErrorResponse:
    properties:
      error:
        example: Invalid request
        type: string
      reqId:
        example: 0b7c0f1e-6a19-4f3f-9a2d-1f9d0f3a6c21
        type: string
    type: object
...
  LoginDiscoverRequest:
    properties:
      email:
        example: alice@acme.com
        maxLength: 320
        type: string
    required:
    - email
    type: object
  LoginDiscoverResponse:
    properties:
      google:
        example: true
        type: boolean
      password:
        example: true
        type: boolean
      sso:
        $ref: '#/definitions/LoginDiscoverSSO'
    type: object
  LoginDiscoverSSO:
    properties:
      connection_id:
        example: 9a8b7c6d-5e4f-3a2b-1c0d-9e8f7a6b5c4d
        type: string
    type: object
...
  LoginResponse:
    properties:
      access_token:
        example: eyJhbGciOiJSUzI1NiIsImtpZCI6ImsxIiwidHlwIjoiYXQrand0In0...
        type: string
      companies:
        items:
          $ref: '#/definitions/CompanyChoice'
        type: array
      expires_in:
        example: 900
        type: integer
      return_to:
        example: /oauth/authorize?client_id=...
        type: string
      status:
        enum:
        - authenticated
        - choose_company
        example: authenticated
        type: string
      token_type:
        example: Bearer
        type: string
    type: object
...
  PasswordLoginRequest:
    properties:
      email:
        example: alice@acme.com
        maxLength: 320
        type: string
      password:
        example: correct-horse-battery-staple
        maxLength: 512
        minLength: 1
        type: string
      return_to:
        description: 'ReturnTo is where App Central goes afterwards: a path on its
          own origin.'
        example: /oauth/authorize?client_id=...
        maxLength: 2048
        type: string
    required:
    - email
    - password
    type: object
...
info:
  contact:
    name: Alora Platform
    url: https://alora.io/support
  description: |-
    App Central: one sign-in for every product of the platform, and the OpenID Provider those products trust.

    ## Two kinds of token
    People sign in once at App Central — with a password, Google, or their company's SSO, as their company's login policy allows. That opens an App Central session (an HttpOnly cookie) and yields an App Central access token (audience `app-central`), the ONLY token the `/api` routes accept. It describes everything about App Central: `scope` (what the person may do here) and `products` (the keys of the apps they may open). A product never sees it: when a product is launched, its backend runs an ordinary OpenID Connect authorization-code flow with PKCE against `/oauth/authorize` and `/oauth/token`, and receives a token for itself alone (audience `product:<key>`, carrying only its own roles).

    ## Scopes
...

    ## Group managers
...

    ## Authorizing requests here
    Sign in with `POST /auth/login/password`, copy `access_token` from the response, then press **Authorize** and paste it. Every `/api` route reads the bearer token from the `Authorization` header only — never from a cookie — and re-checks its session against the database on every request.

    ## Companies
...

    ## Errors
...
  license:
    name: Proprietary
  termsOfService: https://alora.io/terms
  title: App Central API
  version: "2.0"
paths:
...
  /auth/central/logout:
    post:
      description: 'Ends the App Central session and every product login opened under
        it: no product can renew its tokens afterwards. A product access token already
        issued stays valid until it expires, unless the product introspects it. Always
        204.'
      responses:
        "204":
          description: Signed out, or there was nothing to end
        "403":
          description: Cross-site request refused
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Sign out of App Central
      tags:
      - sign-in
  /auth/central/refresh:
    post:
      description: |-
        Rotates the refresh token in the session cookie and returns a new App Central access token. Takes no body. The session's rules are checked again first: the user and company active, the login policy still allowing how the user signed in (a session that no longer qualifies is ended), and the absolute fourteen-day cap.

        Presenting a token that was already rotated away means two parties hold the chain, so the session and every product login under it are revoked. The one exception is a benign concurrent rotation, answered 409 with the cookie left intact.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CentralTokenResponse'
        "401":
          description: Session invalid or expired; cookie cleared
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Cross-site request refused
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Concurrent rotation — retry
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Refresh the App Central session
      tags:
      - sign-in
  /auth/google/callback:
    get:
      description: |-
        The return leg from Google. Always answers 302: to the return path on success, to /login for a choice between companies, or to /login?google_error=<code> on failure.

        The state must match the browser's cookie and still be pending (single use, ten minutes). Google's ID token must verify — RS256 against Google's keys, issuer, audience, expiry and nonce. A linked Google identity signs in its accounts; otherwise a verified address signs in, and links, each live account that holds it. Google never creates an account, and each account's policy must allow Google.
      parameters:
      - description: Must match the browser's pending sign-in
        in: query
        name: state
        required: true
        type: string
      - description: Google's authorization code
        in: query
        name: code
        required: true
        type: string
      responses:
        "302":
          description: Redirect to /login?google_error=invalid_request, state_mismatch,
            state_expired, exchange_failed, email_unverified, account_unavailable
            or server_error
      summary: Google sign-in callback
      tags:
      - sign-in
  /auth/google/start:
    get:
      description: 'A browser navigation: answers 302 to Google''s consent screen,
        with PKCE, a state bound to this browser by an HttpOnly cookie, and a nonce
        bound to the ID token. `return_to` must be a path on App Central; anything
        else is dropped.'
      parameters:
      - description: Where to go after signing in
        in: query
        name: return_to
        type: string
      responses:
        "302":
          description: Redirect to /login?google_error=google_disabled when Google
            sign-in is not configured
      summary: Begin Google sign-in
      tags:
      - sign-in
  /auth/login/choices:
    get:
      description: For the page shown after Google proved accounts in several companies.
        The choice is bound to this browser by an HttpOnly cookie and lasts five minutes.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/CompanyChoices'
        "401":
          description: No pending choice
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: List the companies of a pending account choice
      tags:
      - sign-in
  /auth/login/choose:
    post:
      consumes:
      - application/json
      description: Starts the App Central session in the chosen company. The choice
        is single-use, and the chosen account's login policy is applied again.
      parameters:
      - description: The company
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/ChooseCompanyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/LoginResponse'
        "401":
          description: No pending choice, or the company is not on offer
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Cross-site request refused
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Complete an account choice
      tags:
      - sign-in
  /auth/login/discover:
    post:
      consumes:
      - application/json
      description: 'Answers from the address''s DOMAIN alone: a company that registered
        the domain on an SSO connection gets that connection; a company with that
        verified domain gets what its default policy allows; any other domain gets
        the defaults. Every address at a domain gets the same answer, so it reveals
        nothing about any account — and every sign-in path enforces the user''s own
        policy regardless.'
      parameters:
      - description: The address
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/LoginDiscoverRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/LoginDiscoverResponse'
      summary: Which sign-in methods to offer
      tags:
      - sign-in
  /auth/login/password:
    post:
      consumes:
      - application/json
      description: |-
        Checks the password against every live account that holds the address. When exactly one opens and its company's login policy allows passwords, the App Central session starts: its refresh token is set as the HttpOnly session cookie and an App Central access token is returned. When the password opens accounts in several companies, the answer is `choose_company` with the list — shown only after the password was right — and POST /auth/login/choose completes it.

        No account, a wrong password, and a right password on an account whose policy does not allow passwords all get the same 401, in the same time. Rate limited per minute by ip|email together.
      parameters:
      - description: Credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/PasswordLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/LoginResponse'
        "400":
          description: Malformed body or unknown field
          schema:
            $ref: '#/definitions/ErrorResponse'
        "401":
          description: Invalid email or password
          schema:
            $ref: '#/definitions/ErrorResponse'
        "403":
          description: Cross-site request refused
          schema:
            $ref: '#/definitions/ErrorResponse'
        "415":
          description: Body is not application/json
          schema:
            $ref: '#/definitions/ErrorResponse'
        "429":
          description: Rate limited
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Sign in with a password
      tags:
      - sign-in
...
  /auth/sso/callback:
    get:
      description: The return leg from the identity provider. Always answers 302.
        The ID token must verify against the provider's own keys (RS256), with its
        issuer, audience, authorized party, expiry and nonce. A linked subject signs
        in its account. Otherwise the first sign-in links by address — only at a domain
        registered on the connection, with a verified address (unless the Owner trusts
        the connection's addresses), to an account that already exists. SSO never
        creates an account.
      parameters:
      - description: Must match the browser's pending sign-in
        in: query
        name: state
        required: true
        type: string
      - description: The provider's authorization code
        in: query
        name: code
        required: true
        type: string
      responses:
        "302":
          description: Redirect to /login?sso_error=<code>
      summary: Company SSO sign-in callback
      tags:
      - sign-in
  /auth/sso/start:
    get:
      description: 'A browser navigation: answers 302 to the company''s identity provider,
        with PKCE, state and nonce. The connection is named by `connection_id`, or
        found from the domain of `email`.'
      parameters:
      - description: The SSO connection
        in: query
        name: connection_id
        type: string
      - description: An address at a domain registered on a connection
        in: query
        name: email
        type: string
      - description: Where to go after signing in
        in: query
        name: return_to
        type: string
      responses:
        "302":
          description: Redirect to /login?sso_error=sso_unavailable
      summary: Begin company SSO sign-in
      tags:
      - sign-in
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'An App Central access token from /auth/login/password or /auth/central/refresh.
      Enter it as: Bearer &lt;token&gt;'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...